requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.12.14
colorama==0.4.6
distro==1.9.0
Django==5.1.4
django-rest-api==0.1.5
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.58.1
packaging==24.2
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pytesseract==0.3.13
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
